feat: replace classes to functions
