+++ src/components/ErrorBoundery/types.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react';
+
+export type ErrorBounderyState = {
+  hasError: boolean;
+};
+export type ErrorBounderyProps = {
+  children: ReactNode;
+};
