+++ src/components/MainSection/types.ts
@@ -1,0 +1,21 @@
+export type PersonType = {
+  birth_year: string;
+  created: string;
+  edited: string;
+  eye_color: string;
+  films: string[];
+  gender: string;
+  hair_color: string;
+  height: string;
+  homeworld: string;
+  mass: string;
+  name: string;
+  skin_color: string;
+  species: string;
+  starships: string;
+  url: string;
+};
+
+export type MainSectionProps = {
+  results: PersonType[];
+};
