+++ src/components/SearchField/style.css
@@ -1,0 +1,41 @@
+.search_field {
+  padding: 20px;
+  margin-bottom: 30px;
+}
+.search_form {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.search_container {
+  background: #fff;
+  height: 40px;
+  width: 300px;
+  border-radius: 30px;
+  padding: 10px 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  transition: 0.8s;
+  margin-right: 20px;
+  box-shadow:
+    4px 4px 6px 0 rgba(255, 255, 255, 0.3),
+    -4px -4px 6px 0 rgba(116, 125, 136, 0.2),
+    inset -4px -4px 6px 0 rgba(255, 255, 255, 0.2),
+    inset 4px 4px 6px 0 rgba(0, 0, 0, 0.2);
+}
+.search_container:hover {
+  border: 2px solid #173d3a;
+}
+
+.search_container .search_input {
+  background: transparent;
+  border: none;
+  outline: none;
+  width: 0px;
+  font-weight: 500;
+  font-size: 16px;
+  transition: 0.8s;
+  width: 100%;
+}
