+++ src/components/SearchField/types.ts
@@ -1,0 +1,4 @@
+export type SearchFieldProps = {
+  onSearch: (query: string) => void;
+  value: string;
+};
