+++ src/components/UI/Buttons/style.css
@@ -1,0 +1,20 @@
+.button {
+  background: linear-gradient(180deg, #226cff 0%, #c226ff 100%);
+  color: #fff;
+  border: 0;
+  font-size: 16px;
+  padding: 10px;
+  transition: 0.8s;
+  cursor: pointer;
+}
+.button__rounded {
+  height: 40px;
+  width: 40px;
+  border-radius: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.button:hover {
+  opacity: 0.8;
+}
