+++ src/index.css
@@ -139,6 +139,7 @@ strong {
 .container {
   margin: 0 auto;
   max-width: 1200px;
+  padding: 10px;
 }
 .column {
   display: flex;
@@ -146,184 +147,12 @@ strong {
   align-items: center;
   justify-content: center;
 }
-.search_field {
-  padding: 20px;
-}
-.search_form {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.search_container {
-  background: #fff;
-  height: 40px;
-  width: 300px;
-  border-radius: 30px;
-  padding: 10px 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  transition: 0.8s;
-  margin-right: 20px;
-  box-shadow:
-    4px 4px 6px 0 rgba(255, 255, 255, 0.3),
-    -4px -4px 6px 0 rgba(116, 125, 136, 0.2),
-    inset -4px -4px 6px 0 rgba(255, 255, 255, 0.2),
-    inset 4px 4px 6px 0 rgba(0, 0, 0, 0.2);
-}
-.search_container:hover {
-  border: 2px solid #173d3a;
-}
+
 .logo_image {
   width: 200px;
 }
-.search_container .search_input {
-  background: transparent;
-  border: none;
-  outline: none;
-  width: 0px;
-  font-weight: 500;
-  font-size: 16px;
-  transition: 0.8s;
-  width: 100%;
-}
-.button {
-  background: linear-gradient(180deg, #226cff 0%, #c226ff 100%);
-  color: #fff;
-  border: 0;
-  font-size: 16px;
-  padding: 10px;
-  transition: 0.8s;
-  cursor: pointer;
-}
-.search_button {
-  height: 40px;
-  width: 40px;
-  border-radius: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.button:hover {
-  opacity: 0.8;
-}
-.characters_list {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-}
-.characters_list_item {
-  background-color: #2c3983;
-  border-radius: 10px;
-  padding: 20px;
-  margin-bottom: 20px;
-  box-shadow: 10px 10px 10px #1b2667;
-  width: calc(33% - 10px);
-}
 
-.characters_list_item_avatar {
-  width: 150px;
-  height: 150px;
-  border-radius: 150px;
-  background-color: #1b2667;
-  margin-bottom: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.characters_list_item_avatar img {
-  filter: invert();
-}
-
-.characters_list_item_name {
-  font-size: 20px;
-  font-weight: 600;
-  margin-bottom: 20px;
-}
-.character_detail {
-  width: 100%;
-  margin-bottom: 10px;
-}
 .error_message {
   margin-top: 40px;
   margin-bottom: 20px;
 }
-
-/* loading styles start */
-.loading_container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 20px;
-}
-.ring {
-  position: relative;
-  width: 150px;
-  height: 150px;
-  background: transparent;
-  border: 3px solid #3c3c3c;
-  border-radius: 50%;
-  text-align: center;
-  line-height: 150px;
-  font-family: sans-serif;
-  font-size: 20px;
-  color: #fff000;
-  letter-spacing: 4px;
-  text-transform: uppercase;
-  text-shadow: 0 0 10px #fff000;
-  box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
-}
-.ring:before {
-  content: '';
-  position: absolute;
-  top: -3px;
-  left: -3px;
-  width: 100%;
-  height: 100%;
-  border: 3px solid transparent;
-  border-top: 3px solid #fff000;
-  border-right: 3px solid #fff000;
-  border-radius: 50%;
-  animation: animateC 2s linear infinite;
-}
-span.circle {
-  display: block;
-  position: absolute;
-  top: calc(50% - 2px);
-  left: 50%;
-  width: 50%;
-  height: 4px;
-  background: transparent;
-  transform-origin: left;
-  animation: animate 2s linear infinite;
-}
-span.circle:before {
-  content: '';
-  position: absolute;
-  width: 16px;
-  height: 16px;
-  border-radius: 50%;
-  background: #fff000;
-  top: -6px;
-  right: -8px;
-  box-shadow: 0 0 20px #fff000;
-}
-@keyframes animateC {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-@keyframes animate {
-  0% {
-    transform: rotate(45deg);
-  }
-  100% {
-    transform: rotate(405deg);
-  }
-}
-/* loading styles end */
